Update: Create property api

+++ apps/property/models.py
@@ -136,7 +136,7 @@ class OtherCharges(models.Model):
 class AddOnsPrice(models.Model):
     property = models.ForeignKey(Property, on_delete=models.CASCADE, related_name='add_ons_prices')
     service = models.CharField(max_length=255)
-    price = models.IntegerField(null=True, blank=True)
+    price = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
 
     class Meta:
         verbose_name = 'Add-on Price'
